refactor(headless): split prompt ab runner modules
